// src/types/namefakes.ts
export interface Namefakes {
	name: string;
	address: string;
	latitude: number;
	longitude: number;
	maiden_name: string;
	birth_data: Date;
	phone_h: string;
	phone_w: string;
	email_u: string;
	email_d: string;
	username: string;
	password: string;
	domain: URL;
	useragent: string;
	ipv4: string;
	macaddress: string;
	plasticcard: string;
	cardexpir: string;
	bonus: number;
	company: string;
	color: string;
	uuid: string;
	height: number;
	weight: number;
	blood: string;
	eye: string;
	hair: string;
	pict: string;
	url: URL;
	sport: string;
	ipv4_url: URL;
	email_url: URL;
	domain_url: URL;
}